requirements/production.txt
Django
django-pipeline
dj-database-url
psycopg2
honcho
gunicorn
django-storages-redux
boto
djangorestframework
django-rest-swagger
djangorestframework-jwt
python-social-auth
hashids
Pillow
